[relates #3087] Migrate netty-http module into netty http tests
