+++ integration-tests/http/netty-http/src/main/java/org/apache/camel/quarkus/component/http/netty/NettyHttpResource.java
@@ -18,20 +18,29 @@ package org.apache.camel.quarkus.component.http.netty;
 
 import java.io.ByteArrayInputStream;
 import java.io.IOException;
+import java.net.HttpURLConnection;
+import java.net.InetSocketAddress;
+import java.net.Proxy;
+import java.net.URL;
 import java.nio.charset.StandardCharsets;
 import java.util.Base64;
+import java.util.HashMap;
+import java.util.Map;
 import java.util.zip.GZIPInputStream;
 
+import io.netty.handler.codec.http.FullHttpResponse;
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.ws.rs.Consumes;
 import jakarta.ws.rs.GET;
 import jakarta.ws.rs.POST;
 import jakarta.ws.rs.Path;
+import jakarta.ws.rs.PathParam;
 import jakarta.ws.rs.Produces;
 import jakarta.ws.rs.QueryParam;
 import jakarta.ws.rs.core.MediaType;
 import jakarta.ws.rs.core.Response;
 import org.apache.camel.Exchange;
+import org.apache.camel.component.netty.http.NettyHttpMessage;
 import org.apache.camel.quarkus.component.http.common.AbstractHttpResource;
 import org.apache.camel.util.IOHelper;
 
@@ -118,5 +127,108 @@ public class NettyHttpResource extends AbstractHttpResource {
                 .withHeader(Exchange.HTTP_METHOD, "GET")
                 .send();
         return exchange.getException().getClass().getName();
+    }
+
+    @GET
+    @Path("/getRequest/{method}/{hName}/{hValue}/{body}")
+    public String getRequest(@PathParam("method") String method, @PathParam("hName") String headerName,
+            @PathParam("hValue") String headerValue,
+            @PathParam("body") String body,
+            @QueryParam("test-port") int port) {
+        return producerTemplate.toF("netty-http:http://localhost:%d/request", port)
+                .withHeaders(Map.of(Exchange.HTTP_METHOD, method, headerName, headerValue))
+                .withBody(body)
+                .request(String.class);
+    }
+
+    @GET
+    @Path("/getResponse/{message}")
+    public String getResponse(@PathParam("message") String message, @QueryParam("test-port") int port) {
+        Exchange exchange = producerTemplate.toF("netty-http:http://localhost:%d/response", port)
+                .withBody(message)
+                .send();
+        FullHttpResponse response = exchange.getIn().getBody(NettyHttpMessage.class).getHttpResponse();
+        String received = exchange.getIn().getBody(String.class);
+        return received + ": " + response.status().reasonPhrase() + " " + response.status().code();
+    }
+
+    @GET
+    @Path("/wildcard/{path}")
+    public String wildcard(@PathParam("path") String path, @QueryParam("test-port") int port) {
+        return producerTemplate.toF("netty-http:http://localhost:%d/%s", port, path)
+                .withHeader(Exchange.HTTP_METHOD, "GET")
+                .request(String.class);
+    }
+
+    @GET
+    @Path("/consumer-proxy")
+    public String proxy(@QueryParam("test-port") int port, @QueryParam("proxy-port") int proxyPort) {
+        final Proxy proxy = new Proxy(Proxy.Type.HTTP, new InetSocketAddress("localhost", proxyPort));
+        final String url = "http://localhost:" + port + "/proxy";
+
+        try {
+            final HttpURLConnection connection = (HttpURLConnection) new URL(url).openConnection(proxy);
+            return new String(connection.getInputStream().readAllBytes());
+        } catch (IOException e) {
+            throw new RuntimeException(e);
+        }
+    }
+
+    @GET
+    @Path("/auth/{path}/{user}/{password}")
+    public Response auth(@QueryParam("test-port") int port, @PathParam("path") String path, @PathParam("user") String user,
+            @PathParam("password") String password) {
+        final Exchange exchange = producerTemplate.toF("netty-http:http://localhost:%d/%s", port, path)
+                .withHeaders(getAuthHeaders(user, password))
+                .send();
+
+        return Response.status(exchange.getIn().getHeader(Exchange.HTTP_RESPONSE_CODE, Integer.class)).build();
+    }
+
+    @GET
+    @Path("/rest/{method}")
+    public String rest(@QueryParam("rest-port") int port, @PathParam("method") String method) {
+        return producerTemplate.toF("netty-http:http://localhost:%d/rest", port)
+                .withHeader(Exchange.HTTP_METHOD, method)
+                .request(String.class);
+    }
+
+    @GET
+    @Path("/rest/pojo/{type}")
+    public String restPojo(@QueryParam("rest-port") int port, @PathParam("type") String type) {
+        final String body;
+        final String contentType;
+        if ("json".equals(type)) {
+            body = "{\"firstName\":\"John\", \"lastName\":\"Doe\"}";
+            contentType = "application/json";
+        } else {
+            body = "<user firstName=\"John\" lastName=\"Doe\"/>";
+            contentType = "text/xml";
+        }
+        return producerTemplate.toF("netty-http:http://localhost:%d/rest/%s", port, type)
+                .withBody(body)
+                .withHeaders(Map.of(Exchange.HTTP_METHOD, "POST", "Content-Type", contentType))
+                .request(String.class);
+    }
+
+    @GET
+    @Path("/jaas/{user}/{password}")
+    public Response auth(@QueryParam("test-port") int port, @PathParam("user") String user,
+            @PathParam("password") String password) {
+        final Exchange exchange = producerTemplate
+                .toF("netty-http:http://localhost:%d/jaas", port)
+                .withHeaders(getAuthHeaders(user, password))
+                .send();
+
+        return Response.status(exchange.getIn().getHeader(Exchange.HTTP_RESPONSE_CODE, Integer.class)).build();
+    }
+
+    private Map<String, Object> getAuthHeaders(String user, String password) {
+        Map<String, Object> headers = new HashMap<>();
+        headers.put(Exchange.HTTP_METHOD, "GET");
+        if (!"null".equals(user)) {
+            headers.put("Authorization", "Basic " + Base64.getEncoder().encodeToString((user + ":" + password).getBytes()));
+        }
+        return headers;
     }
 }
