+++ integration-tests/http/netty-http/src/main/java/org/apache/camel/quarkus/component/http/netty/auth/TestAuthenticator.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.quarkus.component.netty.http.auth;
+package org.apache.camel.quarkus.component.http.netty.auth;
 
 import java.security.Principal;
 import java.util.Arrays;
