--- integration-tests/netty-http/src/test/java/org/apache/camel/quarkus/component/netty/http/NettyHttpAuthTest.java
@@ -1,65 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.apache.camel.quarkus.component.netty.http;
-
-import io.quarkus.test.common.QuarkusTestResource;
-import io.quarkus.test.junit.QuarkusTest;
-import io.restassured.RestAssured;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.CsvSource;
-
-@QuarkusTest
-@QuarkusTestResource(NettyHttpTestResource.class)
-public class NettyHttpAuthTest {
-    @ParameterizedTest
-    @CsvSource({
-            "null,null,401",
-            "admin,wrongpass,401",
-            "admin,adminpass,200"
-    })
-    public void testCredentials(String user, String password, int responseCode) {
-        RestAssured
-                .when()
-                .get("/netty/http/auth/{path}/{user}/{password}", "auth?securityConfiguration=#securityConfig", user, password)
-                .then()
-                .statusCode(responseCode);
-    }
-
-    @ParameterizedTest
-    @CsvSource({
-            "admin,admin,adminpass,200",
-            "admin,guest,guestpass,401",
-            "admin,null,null,401",
-            "guest,admin,adminpass,200",
-            "guest,guest,guestpass,200",
-            "guest,null,null,401",
-            "wildcard,admin,adminpass,200",
-            "wildcard,guest,guestpass,200",
-            "wildcard,null,null,401",
-            "public,admin,adminpass,200",
-            "public,guest,guestpass,200",
-            "public,null,null,200",
-    })
-    public void testAcls(String endpoint, String user, String password, int responseCode) {
-        RestAssured
-                .when()
-                .get("/netty/http/auth/{path}/{user}/{password}", "acls/" + endpoint + "?securityConfiguration=#acl" + endpoint,
-                        user, password)
-                .then()
-                .statusCode(responseCode);
-    }
-}
