--- integration-tests/netty-http/src/test/java/org/apache/camel/quarkus/component/netty/http/NettyHttpTestResource.java
@@ -1,37 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.apache.camel.quarkus.component.netty.http;
-
-import java.util.Map;
-import java.util.Objects;
-
-import io.quarkus.test.common.QuarkusTestResourceLifecycleManager;
-import org.apache.camel.quarkus.test.AvailablePortFinder;
-
-public class NettyHttpTestResource implements QuarkusTestResourceLifecycleManager {
-    @Override
-    public Map<String, String> start() {
-        return AvailablePortFinder.reserveNetworkPorts(
-                Objects::toString,
-                "camel.netty-http.port", "camel.netty-http.proxyPort", "camel.netty-http.restPort");
-    }
-
-    @Override
-    public void stop() {
-        AvailablePortFinder.releaseReservedPorts();
-    }
-}
